(Maint) Follow latest gke stable release
So we no longer have to manually update the gke version this commit uses
the google_container_engine_versions resource for determining what the
latest stable release is of a given release channel. This has also been
locked down to the current supported major/minor version with a version
prefix check.