(RE-779) Update kots versions
We are using an older version of kots that consistently fails at
applying our current replicated releases due to annotations being added
to all kubernetes resources regardless of if they are mutable or
immutable.

Later versions of kots no longer add this annotation and we
should use them instead. This commit bumps the version of kots to the
latest version as of this commit.