Add separate Service Account to bastion host
The bastion host was using the GKE cluster's Service Account, which was way too
powerful. This adds a separate SA for the bastion. As we are using OS Login on
the bastion, this SA has no special permissions, ensuring that the user cannot
perform any actions outside their own permissions by impersonating the
instance's SA.

Tested:
[x] I can still SSH into the bastion
[x] I can still renew my credentials for kubectl and perform actions on the
    cluster (like deleting pods, editing deployments)
[x] I can still SSH into Nomad clients from the bastion

SERVER-568 #time 20m #transition in-review