Secure Nomad clients via iptable rules
This adds iptable rules to Nomad clients preventing all docker jobs from
talking to the meta-data service or any internal IPs on the network. I
copied the assumption we make in the GKE k8s module that the internal
network will use a 10.0.0.0/8 block (or part thereof) for the internal
IPs.

I also added an additional firewall rule to allow the rerun-with-ssh
functionality for GKE clusters.