Hotfix - fix external VMs
As some communication between nomad clients, the cluster, and the external VMs is conducted via the public internet, we need to reinstate this firewall rule for all configurations for now. However, we hope to get this fixed on the product level, hence the variables introduced in the original PR remain in place for now.