Use public k8s endpoint by default, disable bastion
This changes the default access pattern from a setup where resources are
private by default and can only be accessed via a bastion host to an
all-public approach. It is the user's responsibility to secure access to
their resources.

The documentation has been updated to reflect this.

SERVER-651 #comment opened PR
SERVER-651 #in-review