Set EKS Nomad Client base AMI to Ubuntu 20.04 Focal
This change sets the base image for the Nomad client to Ubuntu 20.04
Focal. The mechanism in which the image is selected has been altered.

Originally, we used the ubuntu_ami map located in the variables.tf
file located in the EKS root directory. The region would be passed
to the map and the proper AMI was derived from the map.

This change adds `ami.tf` which will perform a lookup against the
Ubuntu AWS AMI list and determine which AMI to use based on the
criteria set in the filter. This value is then made available to the
`ami_id` variable in the `nomad.tf` file.

The `ami_id` was slightly altered to allow the user to provide their
own AMI id. However, it defaults to the auto-selection outlined
above if the variable is not defined.