Set GKE Nomad Client base AMI to Ubuntu 20.04 Focal
This change sets the base image for the Nomad client to Ubuntu 20.04
Focal.

Adds a new variable in the root GKE directory called
`nomad_source_image` which defaults to Ubuntu 20.04 Focal. This allows
the user to easily alter the base image if they need to.

This gets passed to the `source_image` variable sent in to the Nomad
module.