Use separate subnets for different components [EKS]
This commit increases security by ensuring that cluster nodes, Nomad
clients and external VMs do not share subnets in EKS.

The subnets are
distributed so that Nomad clients get half of the subnets (rounded up if
necessary), VMs all go into the same subnet and the cluster nodes get
the remaining subnets. If the Nomad clients are deployed to public
subnets, the all the private subnets can be used for cluster nodes
except one private subnet that is reserved for the VMs.

The Nomad clients now block all access from jobs to any VPC IP, except
the CIDR block for the subnet of the VMs and the VPC DNS server. This
blocks all access to any internal services and IPs from within standard
jobs. It does not work for jobs on external VMs, though.