Add NACL for VM subnet
This adds a Network Access List to the subnet for the external VMs. The
NACL blocks any UDP traffic to other VPC resources and only allows
ephemeral ports for TCP traffic within the VPC (this is still needed for
communicating with vm-service). This should help minimizing the attack
surface further