reduce user-data size
The current iteration of nomad-aws user-data gets really close to the
maximum number of bytes that AWS allows for (it complains at around 21,600
bytes in base64 encoded format).

The cloudinit Terraform provider makes it pretty transparent to compress
and base64 encode user data, it resulted in a size reduction of 50% in
my case.

The use case is if someone forks this module and adds additional
content, they will very quickly hit that limit (which is what happened
to me)